+++ src/App.js
@@ -4,6 +4,44 @@ import logo from './logo.svg';
 import './App.css';
 import Cart from './Cart';
 import CartPage from './CartPage';
+import { useState } from 'react';
+
+// Mock data for demonstration
+const initialCart = [
+  {
+    id: 1,
+    product: {
+      name: 'Product 1',
+      link: '#',
+      imageSrc: 'https://via.placeholder.com/120x120/4caf50/ffffff?text=Product+1',
+      imageAlt: 'Product 1',
+      price: '10.99',
+    },
+    quantity: 4,
+  },
+  {
+    id: 2,
+    product: {
+      name: 'Product 1',
+      link: '#',
+      imageSrc: 'https://via.placeholder.com/120x120/4caf50/ffffff?text=Product+1',
+      imageAlt: 'Product 1',
+      price: '10.99',
+    },
+    quantity: 1,
+  },
+  {
+    id: 3,
+    product: {
+      name: 'Product 2',
+      link: '#',
+      imageSrc: 'https://via.placeholder.com/120x120/2196f3/ffffff?text=Product+2',
+      imageAlt: 'Product 2',
+      price: '20.99',
+    },
+    quantity: 1,
+  },
+];
 
 function Home() {
   return (
@@ -27,7 +65,34 @@ function Home() {
 }
 
 function App() {
-  return <CartPage />;
+  const [cart, setCart] = useState(initialCart);
+
+  const handleQuantityChange = (id, newQuantity) => {
+    setCart(cart =>
+      cart.map(item =>
+        item.id === id
+          ? { ...item, quantity: Math.max(1, Math.min(newQuantity, 99)) }
+          : item
+      )
+    );
+  };
+
+  const handleRemove = id => {
+    setCart(cart => cart.filter(item => item.id !== id));
+  };
+
+  const handleCheckout = () => {
+    alert('Proceeding to checkout!');
+  };
+
+  return (
+    <CartPage
+      cart={cart}
+      onQuantityChange={handleQuantityChange}
+      onRemove={handleRemove}
+      onCheckout={handleCheckout}
+    />
+  );
 }
 
 export default App;
