+++ src/CartPage.js
@@ -1,66 +1,8 @@
-import { useState } from 'react';
+import styles from './CartPage.module.css';
 import CartList from './components/organisms/CartList';
 import CartSidebar from './components/organisms/CartSidebar';
-import styles from './CartPage.module.css';
 
-// Mock data for demonstration
-const initialCart = [
-  {
-    id: 1,
-    product: {
-      name: 'Product 1',
-      link: '#',
-      imageSrc: 'https://via.placeholder.com/120x120/4caf50/ffffff?text=Product+1',
-      imageAlt: 'Product 1',
-      price: '10.99',
-    },
-    quantity: 4,
-  },
-  {
-    id: 2,
-    product: {
-      name: 'Product 1',
-      link: '#',
-      imageSrc: 'https://via.placeholder.com/120x120/4caf50/ffffff?text=Product+1',
-      imageAlt: 'Product 1',
-      price: '10.99',
-    },
-    quantity: 1,
-  },
-  {
-    id: 3,
-    product: {
-      name: 'Product 2',
-      link: '#',
-      imageSrc: 'https://via.placeholder.com/120x120/2196f3/ffffff?text=Product+2',
-      imageAlt: 'Product 2',
-      price: '20.99',
-    },
-    quantity: 1,
-  },
-];
-
-const CartPage = () => {
-  const [cart, setCart] = useState(initialCart);
-
-  const handleQuantityChange = (id, newQuantity) => {
-    setCart(cart =>
-      cart.map(item =>
-        item.id === id
-          ? { ...item, quantity: Math.max(1, Math.min(newQuantity, 99)) }
-          : item
-      )
-    );
-  };
-
-  const handleRemove = id => {
-    setCart(cart => cart.filter(item => item.id !== id));
-  };
-
-  const handleCheckout = () => {
-    alert('Proceeding to checkout!');
-  };
-
+const CartPage = ({ cart, onQuantityChange, onRemove, onCheckout }) => {
   const total = cart
     .reduce((sum, item) => sum + parseFloat(item.product.price) * item.quantity, 0)
     .toFixed(2);
@@ -68,11 +10,11 @@ const CartPage = () => {
   return (
     <main className={styles.main}>
       <div className={styles.layout}>
-        <CartList items={cart} onQuantityChange={handleQuantityChange} onRemove={handleRemove} />
-        <CartSidebar total={total} onCheckout={handleCheckout} />
+        <CartList items={cart} onQuantityChange={onQuantityChange} onRemove={onRemove} />
+        <CartSidebar total={total} onCheckout={onCheckout} />
       </div>
     </main>
   );
 };
 
-export default CartPage; 
+export default CartPage;
